wxGUI/vdigit: category dialog can handle duplicated features (merge from devbr6, https://trac.osgeo.org/grass/changeset/33359, https://trac.osgeo.org/grass/changeset/33360, https://trac.osgeo.org/grass/changeset/33361)
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@33362 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ gui/wxpython/vdigit/driver.h
@@ -174,7 +174,7 @@ public:
 
     std::vector<int> GetSelected(bool);
     std::map<int, std::vector <int> > GetDuplicates();
-    std::vector<int> GetRegionSelected();
+    std::vector<double> GetRegionSelected();
     int SetSelected(std::vector<int>, bool);
     int UnSelect(std::vector<int>);
     std::vector<int> GetSelectedVertex(double, double, double);
